Meilleur composant références, gestion de listes ou objets
