+++ source/sites/publicodes/App.tsx
@@ -13,8 +13,8 @@ import {
 import Tracker, { devTracker } from '../../Tracker'
 import About from './About'
 import Contribution from './Contribution'
-import Ferry from './ferry/Ferry'
-import Carburants from './carburants/Carburants'
+//import Ferry from './ferry/Ferry'
+//import Carburants from './carburants/Carburants'
 import GameOver from './GameOver'
 import Instructions from './Instructions'
 import Landing from './Landing'
@@ -67,8 +67,10 @@ const Router = ({}) => (
 				<Route path="/à-propos" component={About} />
 				<Route path="/vie-privée" component={Privacy} />
 				<Route path="/nouveautés" component={News} />
+				{/*
 				<Route path="/ferry" component={Ferry} />
 				<Route path="/carburants" component={Carburants} />
+				*/}
 				<Route path="/wiki" component={Wiki} />
 				<Route path="/scénarios" component={Scenarios} />
 				<Route path="/crédit-climat-personnel" component={CreditExplanation} />
