+++ source/sites/publicodes/pages/Documentation.tsx
@@ -15,6 +15,8 @@ import Méthode from './Méthode'
 import styled from 'styled-components'
 import Helmet from 'react-helmet'
 import { Markdown } from 'Components/utils/markdown'
+import { Route } from 'react-router'
+import References from './DocumentationReferences'
 
 export default function () {
 	const currentSimulation = useSelector(
@@ -107,99 +109,16 @@ export const DocumentationStyle = styled.div`
 	max-width: 850px;
 	margin: 0 auto;
 	padding: 0 0.6rem;
+	header {
+	}
+	span {
+		background: inherit;
+	}
 	small {
 		background: none !important;
 	}
 
 	div[name='somme'] > div > div:nth-child(2n) {
 		background: var(--darkerColor);
-	}
-`
-
-const referencesImages = {
-	'service-public.fr': '/références-images/marianne.png',
-	'legifrance.gouv.fr': '/références-images/marianne.png',
-	'urssaf.fr': '/références-images/Urssaf.svg',
-	'secu-independants.fr': '/références-images/Urssaf.svg',
-	'gouv.fr': '/références-images/marianne.png',
-	'agirc-arrco.fr': '/références-images/agirc-arrco.png',
-	'pole-emploi.fr': '/références-images/pole-emploi.png',
-	'ladocumentationfrançaise.fr':
-		'/références-images/ladocumentationfrançaise.png',
-	'senat.fr': '/références-images/senat.png',
-	'ameli.fr': '/références-images/ameli.png',
-	'bpifrance-creation.fr': '/références-images/bpi-création.png',
-}
-
-type ReferencesProps = {
-	references: React.ComponentProps<
-		NonNullable<
-			React.ComponentProps<typeof RulePage>['renderers']['References']
-		>
-	>['references']
-}
-
-export function References({ references }: ReferencesProps) {
-	const cleanDomain = (link: string) =>
-		(link.includes('://') ? link.split('/')[2] : link.split('/')[0]).replace(
-			'www.',
-			''
-		)
-
-	return (
-		<StyledReferences>
-			{Object.entries(references).map(([name, link]) => {
-				const domain = cleanDomain(link)
-				return (
-					<li key={name}>
-						<span className="imageWrapper">
-							{Object.keys(referencesImages).includes(domain) && (
-								<img
-									src={
-										referencesImages[domain as keyof typeof referencesImages]
-									}
-									alt={`logo de ${domain}`}
-								/>
-							)}
-						</span>
-						<a href={link} target="_blank" rel="noreferrer">
-							{capitalise0(name)}
-						</a>
-						<span className="ui__ label">{domain}</span>
-					</li>
-				)
-			})}
-		</StyledReferences>
-	)
-}
-
-const StyledReferences = styled.ul`
-	list-style: none;
-	padding: 0;
-	a {
-		flex: 1;
-		min-width: 45%;
-		text-decoration: underline;
-		margin-right: 1rem;
-	}
-	li {
-		margin-bottom: 0.6em;
-		width: 100%;
-		display: flex;
-		align-items: center;
-	}
-	.imageWrapper {
-		width: 4.5rem;
-		height: 3rem;
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		margin-right: 1rem;
-	}
-	img {
-		max-height: 3rem;
-		vertical-align: sub;
-		max-width: 100%;
-		border-radius: 0.3em;
 	}
 `
